chore: improve configurations

+++ app/src/main/java/com/devblack21/rinha/backend/crebito/controller/dto/TransactionRequest.java
@@ -3,7 +3,7 @@ package com.devblack21.rinha.backend.crebito.controller.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 
-public record TransactionRequest(@JsonProperty(value = "valor", required = true) int valor,
+public record TransactionRequest(@JsonProperty(value = "valor", required = true) long valor,
                                  @JsonProperty(value = "tipo", required = true) char tipo,
                                  @JsonProperty(value = "descricao", required = true) String descricao) {
